Add files via upload
Many people travel these days, and the tourist industry is expanding quickly. Therefore, the project's goal is to increase London's tourism industry. This will make it easier for people to find attractions and things to do in London. Users may easily find parks, theaters, and museums in London with the help of this software. In a similar vein, visitors may check each location's opening and closing hours in addition to the entrance price.
The system being implemented for this project is built using the Java programming language. The creation of this program will also make use of well-known design patterns in addition to concepts like OOP methods, objects, classes, generalization, inheritance, etc.